requirements.txt
flask
requests
python-dotenv
paho-mqtt
